Create list item class in separate module
