+++ src/index.js
@@ -1,5 +1,15 @@
 import './style.css';
 import List from './modules/List.js';
+import ListItem from './modules/List item.js';
+
 
 const list1 = new List();
 list1.render();
+console.log(list1.ListObjects);
+
+const listItem1 = new ListItem();
+
+
+
+
+console.log(listItem1)
