+++ src/modules/ListProperties.js
@@ -1,5 +1,5 @@
 export default class ListProperties {
-  constructor(description, completed, index) {
+  constructor(description, completed = false, index) {
     this.description = description;
     this.completed = completed;
     this.index = index;
